Close record select when tabbed-away
If the user tabs away from the search text-field for a record select, then the
dropdown container does not close.  Only a click on the document body will
close the container.  The problem with the container not closing is if the user
tabs over multiple record-select fields, then multiple containers will be
displayed on top of each other.  The container currently opens using a focus
observer, so added a blur observer to close the container when the field is
tabbed away.  There still is a race condition from when the focus event
launchers the Ajax.Updater to fetch the container contents and when it
completes and the container is shown, so add a check to the onComplete to not
show the container if the associated tesxt-field is no longer focused.  The
focused status of the text-field is handled in by the focus/blur observer.