Fix close of record-select container when tabbed away
Using the blur listener to close the container resulted in not being able to
select an item in the container because when the item was clicked it meant the
blur event fired for text-field and closed the container before the click event
for the item fired.  The blur event listener is still needed to set that the
text-field is not focused for the race condition fix for the Ajax.Updater, but
the TAB handling has to move the to onkeypress for the container where it can
close the container and then bubble up to the text-fields to switch the focused
field.