Display record_multi_select_field selections in table instead of list
Record select uses an unordered list with each list item being a selection by
default with the delete (X) link before the selection to remove it from the
list.  The (X) icon was confusing users and the format didn't conform with the
active_scaffold interface.  Changed list to table.  Table layout matches
activescaffold with delete button being a link to the right of the selection
to be deleted. Delete button is plain text instead of (X) icon.